+++ src/serverContext.ts
@@ -29,8 +29,8 @@ import * as ls from "vscode-languageserver-types";
 import { CclsErrorHandler } from "./cclsErrorHandler";
 import { cclsChan, logChan } from './globalContext';
 import { CallHierarchyProvider } from "./hierarchies/callHierarchy";
-import { InactiveRegionsProvider } from "./hierarchies/inactiveRegions";
-import { InheritanceHierarchyProvider } from "./inheritanceHierarchy";
+import { InheritanceHierarchyProvider } from "./hierarchies/inheritanceHierarchy";
+import { InactiveRegionsProvider } from "./inactiveRegions";
 import { PublishSemanticHighlightArgs, SemanticContext, semanticTypes } from "./semantic";
 import { StatusBarIconProvider } from "./statusBarIcon";
 import { ClientConfig, IHierarchyNode } from "./types";
